requirements.txt
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.10.0
filelock==3.9.0
fonttools==4.39.3
idna==2.10
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
onnx==1.13.1
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
Pillow==9.3.0
protobuf==3.20.3
psutil==5.9.4
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
roboflow==1.0.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.19.0
six==1.16.0
sympy==1.11.1
thop==0.1.1.post2209072238
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.61
urllib3==1.26.13
wget==3.2
